as a system admin i want to upload the staff main data as a system i want to generate SUIN ( seasonal unified identifier number ) for the staff added , updated for the received season HAJ-884  , HAJ-888 , HAJ-887

+++ shj-admin-portal-services/src/main/java/com/elm/shj/admin/portal/services/dto/CompanyStaffDto.java
@@ -83,7 +83,7 @@ public class CompanyStaffDto {
     @CellIndex(index = 11)
     private String titleCode;
     @CellIndex(index = 15)
-    private Long season;
+    private int season;
 
     @NullOrNotBlank(min = 10, max = 150)
     @CellIndex(index = 6)
